Overhaul parasite handling
Before:
We tried to ensure that every system defined in a given .asd would
build.  When quicklisp-to-nix-system-info was asked to describe a
parasite (a system that lives in an asd file with a different name),
it would just say "I don't know.  Go look at the host.".  Then, when
quicklisp-to-nix-system-info was handed a host, it would try to
quickload all the parasites.  Dependencies of the parasites were
treated like dependencies of the host.

This breaks when there are dependency cycles between parasites and
their hosts.  For example, simple-date hosts
simple-date/postgres-glue.  simple-date/postgres-glue ultimately
depends on simple-date.  There's no real dependency cycle, here.
simple-date doesn't normally depend on simple-date/postgres-glue, but
we were effectively forcing it to.

After:
Now, we're going to treat parasites like we treat siblings.  Siblings
are systems that are defined in the same quicklisp release.  That is,
one tarball contains multiple asd files.  When generating a nix
package, we move the sibling asd files aside.  If you want to load
that sibling system, there's a different nix package where your
desired system is available.

If you ask for, for example, simple-date, we'll remove
simple-date/postgres-glue from the asd.  If you ask for
simple-date/postgres-glue, we'll remove simple-date from the asd.  One
asd per nix package, and one defsystem per asd.  Now our nix packages
are one-to-one with quicklisp systems!  If quicklisp doesn't see a
dependency, we shouldn't either.

That's the theory, any way.