Introduce a script that generates a nix expression for all quicklisp releases
Reminder: quicklisp-the-package-manager has dists, releases, and
systems.  Most people probably only ever have one dist: the quicklisp
dist.  A dist defines a set of releases.  Releases are just tarballs
of source.  The source of a release includes ASDF systems.

quicklisp-releases-to-nix produces a nix expression that describes all
the releases in the quicklisp dist.  The nix expression it generates
doesn't try to compile the lisp code or establish any sort of
dependencies.  Hopefully, at the end of the day, your typical lisp
programmer won't need to reach for this attr set.

Our goal is to use this attr set as a stepping stone.  We're going to
define a nix function that figures out what releases a given system
depends on and produces a derivation for that system.  As a
convenience, we'll define an attr set with derivations for commonly
used systems.